Feat: Develop overall game system
- Feat: 플레이어 스탯 시스템
- Feat: 씬 전환
- Feat: 포탈 시스템
- Add: 한글  TMP

+++ 2025_SMU_3D_Game_Task/Assets/Scripts/Enemy/EnemySO.cs
@@ -9,7 +9,7 @@ public class EnemySO : ScriptableObject
     [Header("Stats")]
     public float maxHealth;
     public float moveSpeed;
-    public float attackDamage;
+    public int attackDamage;
     public float attackDelayTime;
     public float attackRange;
